+++ src/Arguments/Parser.php
@@ -2,8 +2,8 @@
 namespace Redbox\Cli\Arguments;
 
 
-class Parser {
-
+class Parser
+{
     /**
      * @var \Redbox\Cli\Arguments\Filter $filter;
      */
@@ -33,7 +33,7 @@ class Parser {
      * Set the filter for the parser.
      *
      * @param Filter $filter
-     * @param Array $arguments
+     * @param array $arguments
      */
     public function setFilter(Filter $filter, array $arguments = []) {
         $this->filter    = $filter;
@@ -60,13 +60,10 @@ class Parser {
     public function parse(array $argv = null)
     {
 
-
         $requiredArguments = $this->filter->required();
         list($shortOptions, $longOptions) = $this->buildOptions();
 
-
         $results = getopt($shortOptions, $longOptions);
-
 
         foreach ($this->arguments as $argument) {
             if (isset($results[$argument->prefix])) {
@@ -77,18 +74,14 @@ class Parser {
                 /**
                  * If we set the default value for this argument we also add it to
                  * the result array or it will fail the argument has the option required by mistake.
-                 *
-                 * If you think i should not do this please report it on the boards and explain why.
-                 * (I'm a little in doubt here)
                  */
                 if ($argument->defaultValue) {
                     $this->manager->set($argument->name, $argument->defaultValue);
-                    $this->manager->hasDefaultValue($argument->name, true);
+                    $this->manager->setHasDefaultValue($argument->name, true);
                     $results[$argument->name] = $this->manager->get($argument->name);
                 }
             }
         }
-
 
         foreach ($requiredArguments as $argument) {
             if (isset($results[$argument->prefix]) === false && isset($results[$argument->longPrefix]) === false) {