+++ CHANGELOG.md
@@ -2,8 +2,16 @@
 
 * Updated the CHANGELOG.md to also contain the missing 1.0 changelog entry. fixed #3
 * Added a new line to Redbox\Cli\Arguments::usage() so it would look better. fixed #2 
-* Added Redbox\Cli\Manager::hasDefaultValue() so the user
-
+* Added Redbox\Cli\Manager::hasDefaultValue() so you know if the default value was set #fixed #5
+* Removed the comment about default arguments and if its a good thing in Redbox\Cli\Arguments\Parser::parse()
+* Changed the comments in examples/defaultvalue.php and examples/default.php to mimic the testcase better.
+* Added examples/hasdefaultvalue.php to show the Redbox\Cli\Manager::hasDefaultValue() fature.
+* Redbox\Cli\Arguments::get() now returns false if the argument is unknown.
+* Removed TestBase.php as it was an unused class.
+* Added more tests, also for the Manager class.
+* Cleaned up the code and added more comments.
+* Commented on all functions specially in Manager.php.
+* Fixed all the open issues for the 1.1.1 milestone.
 
 ## 1.1 - 2015-13-12